Checking of the indexes that are on the deletion
There is a possibility that an index that is on a deletion can be selected by the optimizer and we get "index unexpectedly deleted" error. This is caused by the fact that the index still holds in the index tree and will be deleted on commit.
To reproduce this error:
1. Drop the index in the first transaction. Not to do commit.
2. Set the breakpoint at phase 4 in delete_index (dfw.cpp)
3. Make the commit of the first transaction. The debugger will stop at the breakpoint.
4. In the second transaction execute the query with the index.
5. Continue running the server from the breakpoint.

The optimizer first executes the BTR_all, and then the index is locked, but it's already deleted.