Fix bug when uncommitted records may become visible after the database sweep
When the server crash happens, the database might have pointer pages with flags which are not consistent with flags on data pages. If ppg_dp_swept is set for a data page which does not actually have dpg_swept set, then records on this page won't be garbage-collected, and they will become visible after OIT is advanced at the end of the sweep. Checking for dead transactions between OIT and OAT helps to determine if the server has been crashing recently. If it has, we cannot trust ppg_dp_swept flag.