Fix #7627: The size of the database with big records becomes bigger after backup/restore
The problem appears when bulk inserts are used (currently during restore). Without the fix the primary data page is released if it has the dpg_large flag, and the new record is inserted on the next empty page. As a result, each primary page has only one record. Generally the fix solves an issue where per-relation cache of data page numbers is not working for tables with big records.